Create Actionscript SQLite
This is a view for a mobile app using actionscript and adobe air to display data using SQLite. This was used in conjuction with a website I created that generates an XML file dynamically using PHP. A different script stores the XML data into SQLite, allowing the app to work offline.